Move FullnodeController to its correct folder. (#1404)
* Move FullnodeController to its correct folder.

* Move models to controller folder
