+++ src/Stratis.Bitcoin/Controllers/NodeController.cs
@@ -4,7 +4,7 @@ using Stratis.Bitcoin.Base;
 using Stratis.Bitcoin.Builder.Feature;
 using Stratis.Bitcoin.Configuration;
 using Stratis.Bitcoin.Connection;
-using Stratis.Bitcoin.Models;
+using Stratis.Bitcoin.Controllers.Models;
 using Stratis.Bitcoin.P2P.Peer;
 using Stratis.Bitcoin.Utilities;
